refactor(registry): split trust and cache tests
