+++ crates/assay-registry/src/trust_next/mod.rs
@@ -3,3 +3,6 @@ pub(in crate::trust) mod cache;
 pub(in crate::trust) mod decode;
 pub(in crate::trust) mod manifest;
 pub(in crate::trust) mod pinned;
+
+#[cfg(test)]
+mod tests;
